backend/pyproject.toml
[project]
name = "filterapps-backend"
version = "0.1.0"
description = "FilterApps backend — FastAPI + Gemini + Supabase"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "PyJWT[crypto]>=2.9",
    "google-genai>=0.3",
    "supabase>=2.9",
    "pillow>=11.0",
    "numpy>=2.1",
    "httpx>=0.27",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
